Refine extra button row
Each extrabutton row has a distinct ID, to prevent opening other rows than
the one that was clicked.
Also adds a field in the table header to match the added button on each line.
The width of the container (set by bootstrap) is less than the new row width,
which makes the article rows expand beyond the parent table.