Use interface instead of impl. as service type
The userDetailsService field was of type UserDetailsServiceImpl, which
is an implementation of the UserDetailsService interface. This poses a
problem when running controller tests, as we then would like to use
another implementation of the interface.

Changing from the implementation to the interface works both
during normal operation and during testing.